chore: bump Gateway API to 1.4.0 and k8s libs to v0.34.1 (#2451)

+++ controller/gateway/controller_reconciler_utils.go
@@ -1098,7 +1098,7 @@ func setDataPlaneIngressServicePorts(
 		}
 		port := operatorv1beta1.DataPlaneServicePort{
 			Name: name,
-			Port: int32(l.Port),
+			Port: l.Port,
 		}
 		switch l.Protocol {
 		case gatewayv1.HTTPSProtocolType:
